+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/Mapper.java
@@ -26,7 +26,9 @@ import org.apache.cassandra.db.marshal.*;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field.Store;
+import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.search.SortField;
+import org.apache.lucene.util.BytesRef;
 
 import java.nio.ByteBuffer;
 import java.util.List;
@@ -278,6 +280,17 @@ public abstract class Mapper {
     public boolean maps(ColumnDefinition column) {
         String name = column.name.toString();
         return mappedColumns.stream().anyMatch(x -> x.equals(name));
+    }
+
+    void validateTerm(String name, BytesRef term) {
+        int maxSize = IndexWriter.MAX_TERM_LENGTH;
+        int size = term.length;
+        if (size > maxSize) {
+            throw new IndexException("Discarding immense term in field='{}', " +
+                                     "Lucene only allows terms with at most " +
+                                     "{} bytes in length; got {} bytes: {}...",
+                                     name, maxSize, size, term.utf8ToString().substring(0, 10));
+        }
     }
 
     protected MoreObjects.ToStringHelper toStringHelper(Object self) {