+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 Merged from 3.0.7:
 * Fix parsing of data with CQL "date" type (#158)
+* Fix handling of immense term insertion mapping error
 
 ## 3.7.0 (20 June 2016)